+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.11.0] - TBD
 ### Added
-- Tree Helper: Pruning tree to allow pruning by `prune_path` and `max_depth.`
+- Tree Helper: Pruning tree to allow pruning by `prune_path` and `max_depth`.
+- Tree Plot: Implement Enhanced Reingold Tilford Algorithm to retrieve (x, y) coordinates for a tree structure.
 ### Fixed
 - Utility Iterator: Relax type hinting using TypeVar.